refactor: move decrypted message structs to a unified module
# Conflicts:
#	crypto-ffi/src/generic/mod.rs
