+++ crypto-ffi/src/crl.rs
@@ -2,10 +2,19 @@
 use wasm_bindgen::prelude::*;
 
 #[derive(Debug, Clone, derive_more::From, derive_more::Into)]
+#[cfg_attr(target_family = "wasm", wasm_bindgen, derive(serde::Serialize, serde::Deserialize))]
 pub struct NewCrlDistributionPoints(Option<Vec<String>>);
 
 #[cfg(not(target_family = "wasm"))]
 uniffi::custom_newtype!(NewCrlDistributionPoints, Option<Vec<String>>);
+
+impl From<core_crypto::e2e_identity::NewCrlDistributionPoints> for NewCrlDistributionPoints {
+    fn from(value: core_crypto::e2e_identity::NewCrlDistributionPoints) -> Self {
+        let value = value.into_iter().collect::<Vec<_>>();
+        let value = (!value.is_empty()).then_some(value);
+        Self(value)
+    }
+}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(target_family = "wasm", wasm_bindgen, derive(serde::Serialize, serde::Deserialize))]
