+++ crypto-ffi/src/wasm/mod.rs
@@ -26,235 +26,9 @@ use wasm_bindgen::{JsCast, prelude::*};
 use wasm_bindgen_futures::future_to_promise;
 
 use crate::{
-    Ciphersuite, CommitBundle, CoreCryptoError, CredentialType, FfiClientId, InternalError, MlsError, ProposalBundle,
-    WasmCryptoResult, lower_ciphersuites,
+    Ciphersuite, CommitBundle, CoreCryptoError, CredentialType, FfiClientId, InternalError, MlsError, WasmCryptoResult,
+    lower_ciphersuites,
 };
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-/// see [core_crypto::prelude::decrypt::MlsConversationDecryptMessage]
-pub struct DecryptedMessage {
-    message: Option<Vec<u8>>,
-    proposals: Vec<ProposalBundle>,
-    /// It is set to false if ingesting this MLS message has resulted in the client being removed from the group (i.e. a Remove commit)
-    is_active: bool,
-    /// Commit delay hint (in milliseconds) to prevent clients from hammering the server with epoch changes
-    commit_delay: Option<u32>,
-    sender_client_id: Option<Vec<u8>>,
-    /// true when the decrypted message resulted in an epoch change i.e. it was a commit
-    has_epoch_changed: bool,
-    identity: WireIdentity,
-    buffered_messages: Option<Vec<BufferedDecryptedMessage>>,
-    /// New CRL Distribution of members of this group
-    crl_new_distribution_points: Option<Vec<String>>,
-}
-
-impl TryFrom<MlsConversationDecryptMessage> for DecryptedMessage {
-    type Error = CoreCryptoError;
-
-    fn try_from(from: MlsConversationDecryptMessage) -> Result<Self, Self::Error> {
-        let proposals = from
-            .proposals
-            .into_iter()
-            .map(ProposalBundle::try_from)
-            .collect::<WasmCryptoResult<Vec<_>>>()?;
-
-        let buffered_messages = if let Some(bm) = from.buffered_messages {
-            let bm = bm
-                .into_iter()
-                .map(TryInto::try_into)
-                .collect::<WasmCryptoResult<Vec<_>>>()?;
-            Some(bm)
-        } else {
-            None
-        };
-
-        let commit_delay = from
-            .delay
-            .map(TryInto::try_into)
-            .transpose()
-            .map_err(InternalError::generic())?;
-
-        #[expect(deprecated)]
-        Ok(Self {
-            message: from.app_msg,
-            proposals,
-            is_active: from.is_active,
-            commit_delay,
-            sender_client_id: from.sender_client_id.map(ClientId::into),
-            has_epoch_changed: from.has_epoch_changed,
-            identity: from.identity.into(),
-            buffered_messages,
-            crl_new_distribution_points: from.crl_new_distribution_points.into(),
-        })
-    }
-}
-
-#[wasm_bindgen]
-impl DecryptedMessage {
-    #[wasm_bindgen(getter)]
-    pub fn message(&self) -> JsValue {
-        if let Some(message) = &self.message {
-            Uint8Array::from(message.as_slice()).into()
-        } else {
-            JsValue::NULL
-        }
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn proposals(&self) -> js_sys::Array {
-        self.proposals
-            .iter()
-            .cloned()
-            .map(JsValue::from)
-            .collect::<js_sys::Array>()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn is_active(&self) -> bool {
-        self.is_active
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn commit_delay(&self) -> Option<u32> {
-        self.commit_delay
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn sender_client_id(&self) -> JsValue {
-        if let Some(cid) = &self.sender_client_id {
-            Uint8Array::from(cid.as_slice()).into()
-        } else {
-            JsValue::NULL
-        }
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn has_epoch_changed(&self) -> bool {
-        self.has_epoch_changed
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn identity(&self) -> WireIdentity {
-        self.identity.clone()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn buffered_messages(&self) -> Option<js_sys::Array> {
-        self.buffered_messages
-            .clone()
-            .map(|bm| bm.iter().cloned().map(JsValue::from).collect::<js_sys::Array>())
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn crl_new_distribution_points(&self) -> Option<js_sys::Array> {
-        self.crl_new_distribution_points
-            .clone()
-            .map(|crl_dp| crl_dp.iter().cloned().map(JsValue::from).collect::<js_sys::Array>())
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-/// to avoid recursion
-pub struct BufferedDecryptedMessage {
-    message: Option<Vec<u8>>,
-    proposals: Vec<ProposalBundle>,
-    is_active: bool,
-    commit_delay: Option<u32>,
-    sender_client_id: Option<Vec<u8>>,
-    has_epoch_changed: bool,
-    identity: WireIdentity,
-    /// New CRL Distribution of members of this group
-    crl_new_distribution_points: Option<Vec<String>>,
-}
-
-impl TryFrom<MlsBufferedConversationDecryptMessage> for BufferedDecryptedMessage {
-    type Error = CoreCryptoError;
-
-    fn try_from(from: MlsBufferedConversationDecryptMessage) -> Result<Self, Self::Error> {
-        let proposals = from
-            .proposals
-            .into_iter()
-            .map(TryInto::try_into)
-            .collect::<WasmCryptoResult<Vec<_>>>()?;
-
-        let commit_delay = from
-            .delay
-            .map(TryInto::try_into)
-            .transpose()
-            .map_err(InternalError::generic())?;
-
-        #[expect(deprecated)]
-        Ok(Self {
-            message: from.app_msg,
-            proposals,
-            is_active: from.is_active,
-            commit_delay,
-            sender_client_id: from.sender_client_id.map(ClientId::into),
-            has_epoch_changed: from.has_epoch_changed,
-            identity: from.identity.into(),
-            crl_new_distribution_points: from.crl_new_distribution_points.into(),
-        })
-    }
-}
-
-#[wasm_bindgen]
-impl BufferedDecryptedMessage {
-    #[wasm_bindgen(getter)]
-    pub fn message(&self) -> JsValue {
-        if let Some(message) = &self.message {
-            Uint8Array::from(message.as_slice()).into()
-        } else {
-            JsValue::NULL
-        }
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn proposals(&self) -> js_sys::Array {
-        self.proposals
-            .iter()
-            .cloned()
-            .map(JsValue::from)
-            .collect::<js_sys::Array>()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn is_active(&self) -> bool {
-        self.is_active
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn commit_delay(&self) -> Option<u32> {
-        self.commit_delay
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn sender_client_id(&self) -> JsValue {
-        if let Some(cid) = &self.sender_client_id {
-            Uint8Array::from(cid.as_slice()).into()
-        } else {
-            JsValue::NULL
-        }
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn has_epoch_changed(&self) -> bool {
-        self.has_epoch_changed
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn identity(&self) -> WireIdentity {
-        self.identity.clone()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn crl_new_distribution_points(&self) -> Option<js_sys::Array> {
-        self.crl_new_distribution_points
-            .clone()
-            .map(|crl_dp| crl_dp.iter().cloned().map(JsValue::from).collect::<js_sys::Array>())
-    }
-}
 
 #[wasm_bindgen(getter_with_clone)]
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
